Introduce T993210 to release lchan + subscr_conn if MSC fails to respond
The GSM specs don't have an explicit timer for this, so let's introdcue
a custom timer (hence starting with 99).

This timeout catches the following situation:
* we send a SCCP CR with COMPL_L3_INFO from the MS to the MSC,
* the MSC doesn't respond (e.g. SCCP routing failure, program down, ...)

The MS is supposed to timeout with T3210, 3220 or 3230.  But the BSC
shouldn't trust the MS but have some timer on its own.

SCCP would have a timer T(conn est), but that one is specified to be
1-2min and hence rather long.

Closes: OS#2775
Change-Id: Ie11d7d06353ba1b1e2fab6763dd7b032ce8a5d2c