handover_test.c: add the DMSC logging category WIP
The new gscon FSM logs to DMSC. Un-mute this logging in handover_test.c.

DISCLAIMER: it's not set in stone that the gscon will remain in the DMSC
logging category, it's just the current dev status-quo. If it moves to a
different category, handover_test.c needs to adjust to it.

Change-Id: I9713a3743c8fb9e7ba6f9c5925311ead8386a09f