abis_rsl: properly handle RSL Release Ind from MS
If we receive a Release Ind from the MS, we so far just drop it on the floor,
assuming that some other part of the BSC code invoked something that it is
already taking care of -- which is of course not the case.

If we receive a Release Ind for an lchan that is not in state LCHAN_S_REL_REQ,
signal the MSC with a Clear Request which will lead to a graceful teardown of
the entire connection.

If there should be no conn, immediately invoke the channel release and ignore
BSSAP.

Related: ttcn3 test TC_chan_rel_rll_rel_ind(),
         see I737be141b69a250eb6eb38007f8042981c1a31cf (osmo-ttcn3-hacks)
Change-Id: I0f8c9c4e6b6850b15c70250fd3f88bdf75f9accf