SCCPlite Assignment Complete: include Speech Codec (Chosen)
Compose the Speech Codec (Chosen) IE not only for AoIP, but also for SCCPlite:
place the code that assigns the codec to sc_ptr outside of the gscon_is_aoip()
if scope.

This way the MSC is told the chosen speech codec, which is mandatory for IP
based user plane, and was missing until now.

Related: OS#3528
Change-Id: Ibedade8d71a7994d25a63bc2faa2a24a10bfffa1