handover: skip legacy neighbors if explicit neighbors exist
We ignore the legacy neighbor-list configuration in the presence of explicit
neighbor mappings -- so far this applied only to local-BSS neighbors. Also skip
the neighbor-list configuration in the presence of remote-BSS neighbors.

At first, I was not entirely sure whether this patch is necessary, but it
serves clarity. Consider: it might make sense to a user to use legacy
'neighbor-list' configuration for local-BSS cells, and only add some remote-BSS
cells using the new 'neighbor' config item; from that POV it makes sense to
still heed 'neighbor-list' even if remote-BSS neighbors are configured, and to
only ignore 'neighbor-list' if there are actual local-BSS neighbor mappings.
However, it is not always immediately obvious whether a given neighbor mapping
refers to a local-BSS or a remote-BSS neighbor. So, let's not convolute the new
and old configuration schemes in weird ways: rather completely ignore a cell's
legacy 'neighbor-list' config as soon as any new 'neighbor' mapping exists on
it.

Change-Id: I3ae22552db521ea640d4343078acb36048119ed2